test: isolate runtime identity service coverage
